Made all cogs folders so you can have more that one file for a cog
