+++ mawabot/cogs/guild/__init__.py
@@ -1,0 +1,6 @@
+from .core import Guild
+
+def setup(bot):
+    ''' Setup function to add cog to bot '''
+    cog = Guild(bot)
+    bot.add_cog(cog)
