+++ mawabot/cogs/info/__init__.py
@@ -1,0 +1,6 @@
+from .core import Info
+
+def setup(bot):
+    ''' Setup function to add cog to bot '''
+    cog = Info(bot)
+    bot.add_cog(cog)
