+++ mawabot/cogs/programming/__init__.py
@@ -1,0 +1,6 @@
+from .core import Programming
+
+def setup(bot):
+    ''' Setup function to add cog to bot '''
+    cog = Programming(bot)
+    bot.add_cog(cog)
