+++ mawabot/cogs/text/core.py
@@ -110,8 +110,3 @@ class Text:
         words.append(last)
 
         await ctx.message.edit(content='\n'.join(words))
-
-def setup(bot):
-    ''' Setup function to add cog to bot '''
-    cog = Text(bot)
-    bot.add_cog(cog)
